Fix compose preview broken because BondedBlePeripherals had a reference to BleManager. Created a Fake version for previews
